size and non-size done
